fix: code review issue
