+++ .env
@@ -1,0 +1,1 @@
+API_CATS_KEY=live_Cek4sP7B5064Il5Nsv9JsIrvjXSDpdGkusCv2TRxsYaZcgdVUyLysAftjm0QJoc7
