+++ src/api/index.tsx
@@ -3,7 +3,7 @@ import axios from 'axios';
 const apiCats = axios.create({
   baseURL: 'https://api.thecatapi.com/v1/',
   headers: {
-    'x-api-key': 'live_Cek4sP7B5064Il5Nsv9JsIrvjXSDpdGkusCv2TRxsYaZcgdVUyLysAftjm0QJoc7',
+    'x-api-key': process.env.APP_CATS_KEY || 'DEMO-API-KEY',
   },
 });
