+++ src/app/page.tsx
@@ -4,10 +4,11 @@ import { useHomeModel } from '@/templates/Home/model';
 import { HomeView } from '@/templates/Home/view';
 
 export default function HomePage() {
-  const { activeCategory, cats, handleCategoryChange, loading } = useHomeModel();
+  const { activeCategory, cats, handleCategoryChange, loading, categories } = useHomeModel();
 
   return (
     <HomeView
+      categories={categories}
       activeCategory={activeCategory}
       cats={cats}
       handleCategoryChange={handleCategoryChange}
